Move the leases_current matcher because we have a whole module for those ...
