+++ src/_zkapauthorizer/tests/test_lease_maintenance.py
@@ -37,7 +37,6 @@ from testtools.matchers import (
     HasLength,
     MatchesAll,
     AllMatch,
-    GreaterThan,
     AfterPreprocessing,
 )
 from testtools.twistedsupport import (
@@ -83,6 +82,7 @@ from ..foolscap import (
 from .matchers import (
     Provides,
     between,
+    leases_current,
 )
 from .strategies import (
     storage_indexes,
@@ -469,31 +469,3 @@ class MaintainLeasesFromRootTests(TestCase):
                 min_lease_remaining,
             ))
         )
-
-
-def leases_current(relevant_storage_indexes, now, min_lease_remaining):
-    """
-    Return a matcher on a ``DummyStorageServer`` instance which matches
-    servers for which the leases on the given storage indexes do not expire
-    before ``min_lease_remaining``.
-    """
-    return AfterPreprocessing(
-        # Get share stats for storage indexes we should have
-        # visited and maintained.
-        lambda storage_server: list(
-            stat
-            for (storage_index, stat)
-            in storage_server.buckets.items()
-            if storage_index in relevant_storage_indexes
-        ),
-        AllMatch(
-            AfterPreprocessing(
-                # Lease expiration for anything visited must be
-                # further in the future than min_lease_remaining,
-                # either because it had time left or because we
-                # renewed it.
-                lambda share_stat: datetime.utcfromtimestamp(share_stat.lease_expiration),
-                GreaterThan(now + min_lease_remaining),
-            ),
-        ),
-    )
